requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
psychometric-tests==0.1.1
psycopg2==2.9.9
pyarrow==15.0.0
pylance==0.11.0
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
shiboken6==6.7.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
